server_code/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SocketIO==2.9.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
Pillow==5.1.0
pkg-resources==0.0.0
pymongo==3.6.1
python-engineio==2.0.4
python-socketio==1.9.0
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.47.0
Werkzeug==0.14.1
WTForms==2.1
